Add files via upload
The HTML is close to being ready to be broken into handlebars. The CSS is close but I still need to move the items on the right into a flex box and the new listings into a flex box so that the empty space at the top is filled. Unless one of you know a way to do that with out putting them in flex boxes. 
The index.html still has the / remove infront of the styly.css and index.js so that I can run is with out the server running.